Refactor + remove redundant rollup setup
