+++ .gitignore
@@ -1,4 +1,3 @@
 .DS_Store
 node_modules
-test_en_*
 dist
